requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
flask>=2.3.0
azure-cognitiveservices-speech>=1.34.0